Make dynamic views

+++ MainApp/Controllers/MailBoxController.cs
@@ -49,36 +49,35 @@ namespace JanuszMail.Controllers
             if (providerParams.Any())
             {
                 _providerParams = providerParams.First();
-                var result = _provider.Connect(_providerParams);
+                HttpStatusCode result;
+                result = _provider.Connect(_providerParams);
                 return result == HttpStatusCode.OK;
             }
             return false;
         }
 
         // GET: MailBox
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(string replyTo = null)
         {
             var connectionStatus = await ConnectToProvider();
             if (!connectionStatus)
             {
-                TempData["ErrorMessage"] = "You have no provider selected. Go to Manage and add provider.";
+                TempData["ErrorMessage"] = "Cannot connect to provider, make sure you added correct provider";
                 return View("Error");
             }
-
-            //TODO add a smart way of determining folder parameter, one that would at least ignore case
-            return RedirectToAction(nameof(ShowMails), new { folder = "INBOX", sortOrder = "dateDesc" });
+            return View(model: replyTo);
         }
+
         public async Task<IActionResult> ShowMails(int? page, int? pageSize, string folder, string sortOrder, string subject, string sender)
         {
             //Should return partial view with PagedList of MailMessages that matches to given params
             //When there is no matching messages then should return partial view with error message
             ViewBag.Folder = folder;
-            ViewBag.ReturnUrl = Url.Action("ShowMails", new { page = page, pageSize = pageSize, folder = folder, sortOrder = sortOrder, sender = sender, subject = subject });
             var connectionStatus = await ConnectToProvider();
             if (!connectionStatus)
             {
                 TempData["ErrorMessage"] = "Something wrong with connection";
-                return View("Error");
+                return null;
             }
             int currentPage = page ?? 1;
             int currentPageSize = pageSize ?? 25;
@@ -88,7 +87,7 @@ namespace JanuszMail.Controllers
             if (!httpStatusCode.Equals(HttpStatusCode.OK))
             {
                 TempData["ErrorMessage"] = "Downloading messages from server failed";
-                return View("Error");
+                return null;
             }
 
             var mails = mailsTuple.Item1.AsQueryable();
@@ -99,7 +98,7 @@ namespace JanuszMail.Controllers
             }
             if (!String.IsNullOrEmpty(sender))
             {
-                mails = mails.Where(mail => mail.Sender.Address.Contains(sender) || mail.Sender.Name.Contains(sender));
+                mails = mails.Where(mail => mail.SenderEmail.Contains(sender) || mail.SenderName.Contains(sender));
             }
 
             ViewBag.DateSortParam = String.IsNullOrEmpty(sortOrder) ? "dateAsc" : ViewBag.DateSortParam;
@@ -109,18 +108,9 @@ namespace JanuszMail.Controllers
             ViewBag.CurrectSortOrder = sortOrder;
 
             currentFolder = folder;
-            List<Mail> myMails = mails.ToList();
-            var results = new StaticPagedList<Mail>(myMails, currentPage, currentPageSize, _provider.GetFolder(currentFolder).Count);
-            if (!results.Any())
-            {
-                TempData["ErrorMessage"] = "No messages matching criteria";
-                return View("Error");
-            }
-            else
-            {
-                //_provider.DownloadAttachment("Duda.jpg", myMails.First().ID, currentFolder);
-                return View(results);
-            }
+            List<MailHeader> headers = mails.Select<Mail, MailHeader>(x => x).ToList();
+            var results = new StaticPagedList<MailHeader>(headers, currentPage, currentPageSize, _provider.GetFolder(currentFolder).Count);
+            return PartialView("_ShowMails", results);
         }
 
         public async Task<IActionResult> Details(uint? id, string folder)
@@ -129,7 +119,7 @@ namespace JanuszMail.Controllers
             if (id == null)
             {
                 TempData["ErrorMessage"] = "Could not open the message";
-                return View("Error");
+                return PartialView("Error");
             }
             if (String.IsNullOrEmpty(folder))
             {
@@ -141,7 +131,7 @@ namespace JanuszMail.Controllers
             if (!connectionStatus)
             {
                 TempData["ErrorMessage"] = "You have no provider selected. Go to Manage and add provider.";
-                return View("Error");
+                return PartialView("Error");
             }
 
             var tuple = await Task.Run(() => { return _provider.GetMailFromFolder(new UniqueId(ID), folder); });
@@ -152,39 +142,28 @@ namespace JanuszMail.Controllers
             if (!httpStatusCode.Equals(HttpStatusCode.OK))
             {
                 TempData["ErrorMessage"] = "Could not open the requested e-mail";
-                return View("Error");
+                return PartialView("Error");
             }
             else
             {
                 ViewBag.Folder = folder;
-                ViewBag.ReturnUrlFailing = Url.Action("Details", new { id = mail.ID.Id, folder = folder });
-                ViewBag.ReturnUrlPassing = Url.Action("ShowMails", new { folder = folder });
                 await markReadTask;
-                return View(mail);
+                return PartialView("_Details", model: mail);
             }
         }
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Delete(uint? id, string folder, string returnUrlPassing, string returnUrlFailing)
+        public async Task<bool> Delete(uint? id, string folder)
         {
             var connectionStatus = await ConnectToProvider();
             if (!connectionStatus)
             {
-                TempData["ErrorMessage"] = "You have no provider selected. Go to Manage and add provider.";
-                return View("Error");
+                return false;
             }
 
-            if (id == null)
+            if (id == null || folder == null)
             {
-                if (folder == null || folder.Length == 0)
-                {
-                    return Redirect(Url.Action("Index"));
-                }
-                else
-                {
-                    TempData["ErrorMessage"] = "Message or value not specified";
-                    return Redirect(returnUrlFailing);
-                }
+                return false;
             }
             uint ID = id ?? 0; //otherwise it won't compile
             if (String.IsNullOrEmpty(folder))
@@ -192,15 +171,7 @@ namespace JanuszMail.Controllers
                 folder = "Inbox";
             }
             var statusCode = await Task.Run(() => { return _provider.RemoveEmail(new UniqueId(ID), folder); });
-            if (statusCode.Equals(HttpStatusCode.OK))
-            {
-                return Redirect(returnUrlPassing);
-            }
-            else
-            {
-                TempData["ErrorMessage"] = "Could not access e-mail server";
-                return Redirect(returnUrlFailing);
-            }
+            return statusCode.Equals(HttpStatusCode.OK);
         }
 
         // GET: MailBox/Create
@@ -210,7 +181,7 @@ namespace JanuszMail.Controllers
             if (!connectionStatus)
             {
                 TempData["ErrorMessage"] = "You have no provider selected. Go to Manage and add provider.";
-                return View("Error");
+                return PartialView("Error");
             }
             var getSignatureTask = Task.Run(async () =>
             {
@@ -227,7 +198,7 @@ namespace JanuszMail.Controllers
                 mail.Recipient = replyTo;
             }
             mail.Body = await getSignatureTask;
-            return View(mail);
+            return PartialView("_Send", mail);
         }
 
         private async Task<MimeMessage> ConstructMimeMessage(Mail mail)
@@ -237,21 +208,21 @@ namespace JanuszMail.Controllers
             mimeMessage.From.Add(new MailboxAddress(_providerParams.EmailAddress));
             mimeMessage.To.Add(new MailboxAddress(mail.Recipient));
             mimeMessage.Subject = mail.Subject;
-
-            foreach (var attachment in mail.Attachments)
-            {
-                var tempPath = Path.GetTempPath();
-                if (attachment.Length > 0)
+            if (mail.Attachments != null)
+                foreach (var attachment in mail.Attachments)
                 {
-                    var filePath = Path.Combine(tempPath, attachment.FileName);
-                    using (var stream = new FileStream(filePath, FileMode.Create))
+                    var tempPath = Path.GetTempPath();
+                    if (attachment.Length > 0)
                     {
-                        await attachment.CopyToAsync(stream);
+                        var filePath = Path.Combine(tempPath, attachment.FileName);
+                        using (var stream = new FileStream(filePath, FileMode.Create))
+                        {
+                            await attachment.CopyToAsync(stream);
+                        }
+                        builder.Attachments.Add(filePath);
+                        System.IO.File.Delete(filePath);
                     }
-                    builder.Attachments.Add(filePath);
-                    System.IO.File.Delete(filePath);
                 }
-            }
             builder.HtmlBody = mail.Body;
             mimeMessage.Body = builder.ToMessageBody();
             return mimeMessage;
@@ -260,15 +231,14 @@ namespace JanuszMail.Controllers
         // POST: MailBox/Create
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost, ActionName("Send")]
+        [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Send([Bind("ID,Recipient,Subject,Attachments,Body")] Mail mail)
+        public async Task<bool> Send([Bind("ID,Recipient,Subject,Attachments,Body")] Mail mail)
         {
             var connectionStatus = await ConnectToProvider();
             if (!connectionStatus)
             {
-                TempData["ErrorMessage"] = "You have no provider selected. Go to Manage and add provider.";
-                return View("Error");
+                return false;
             }
             if (ModelState.IsValid)
             {
@@ -277,15 +247,10 @@ namespace JanuszMail.Controllers
                 HttpStatusCode httpStatusCode = await Task.Run(() => { return _provider.SendEmail(mail.mimeMessage); });
                 if (httpStatusCode.Equals(HttpStatusCode.OK))
                 {
-                    return RedirectToAction(nameof(Index));
-                }
-                else
-                {
-                    TempData["ErrorMessage"] = "Could not send the message";
-                    return View(mail);
+                    return true;
                 }
             }
-            return View(mail);
+            return false;
         }
         [HttpGet, ActionName("DownloadAttachment")]
         public async Task<ActionResult> DownloadAttachment(int id, string fileName, string folder)
@@ -306,102 +271,57 @@ namespace JanuszMail.Controllers
         }
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> MarkRead(uint? id, string folder, string returnUrlPassing, string returnUrlFailing)
+        public async Task<bool> MarkRead(uint? id, string folder)
         {
             var connectionStatus = await ConnectToProvider();
             if (!connectionStatus)
             {
-                TempData["ErrorMessage"] = "You have no provider selected. Go to Manage and add provider.";
-                return View("Error");
+                return false;
             }
 
-            if (id == null)
+            if (id == null || folder == null)
             {
-                if (folder == null || folder.Length == 0)
-                    return Redirect(Url.Action("Index"));
-                else
-                {
-                    TempData["ErrorMessage"] = "Message or value not specified";
-                    return Redirect(returnUrlFailing);
-                }
+                return false;
             }
             HttpStatusCode result = await Task.Run(() => { return _provider.MarkEmailAsRead(new UniqueId(id ?? 0), folder); });
-            if (result == HttpStatusCode.OK)
-            {
-                return Redirect(returnUrlPassing);
-            }
-            else
-            {
-                TempData["ErrorMessage"] = "Something wrong with marking message";
-                return Redirect(returnUrlFailing);
-            }
+            return (result == HttpStatusCode.OK);
         }
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> MarkUnread(uint? id, string folder, string returnUrlPassing, string returnUrlFailing)
+        public async Task<bool> MarkUnread(uint? id, string folder)
         {
             var connectionStatus = await ConnectToProvider();
             if (!connectionStatus)
             {
-                TempData["ErrorMessage"] = "You have no provider selected. Go to Manage and add provider.";
-                return View("Error");
+                return false;
             }
 
-            if (id == null)
+            if (id == null || folder == null)
             {
-                if (folder == null || folder.Length == 0)
-                    return Redirect(Url.Action("Index"));
-                else
-                {
-                    TempData["ErrorMessage"] = "Message or value not specified";
-                    return Redirect(returnUrlFailing);
-                }
+                return false;
             }
             HttpStatusCode result = await Task.Run(() => { return _provider.MarkEmailAsUnread(new UniqueId(id ?? 0), folder); });
 
-            if (result == HttpStatusCode.OK)
-            {
-                return Redirect(returnUrlPassing);
-            }
-            else
-            {
-                TempData["ErrorMessage"] = "Something wrong with marking message";
-                return Redirect(returnUrlFailing);
-            }
+            return (result == HttpStatusCode.OK);
         }
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> MoveToFolder(uint? id, string folder, string destFolder, string returnUrlPassing, string returnUrlFailing)
+        public async Task<bool> MoveToFolder(uint? id, string folder, string destFolder)
         {
             var connectionStatus = await ConnectToProvider();
             if (!connectionStatus)
             {
-                TempData["ErrorMessage"] = "You have no provider selected. Go to Manage and add provider.";
-                return View("Error");
+                return false;
             }
 
-            if (id == null)
+            if (id == null || folder == null)
             {
-                if (folder == null || folder.Length == 0)
-                    return Redirect(Url.Action("Index"));
-                else
-                {
-                    TempData["ErrorMessage"] = "Message or value not specified";
-                    return Redirect(returnUrlFailing);
-                }
+                return false;
             }
             HttpStatusCode result = await Task.Run(() => { return _provider.MoveEmailToFolder(new UniqueId(id ?? 0), folder, destFolder); });
 
-            if (result == HttpStatusCode.OK)
-            {
-                return Redirect(returnUrlPassing);
-            }
-            else
-            {
-                TempData["ErrorMessage"] = "Something wrong with marking message";
-                return Redirect(returnUrlFailing);
-            }
+            return (result == HttpStatusCode.OK);
         }
         private readonly IProvider _provider;
         private readonly UserManager<ApplicationUser> _userManager;
